backend/requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncpg==0.31.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.6
distro==1.9.0
docker==7.1.0
ecdsa==0.19.2
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.135.3
fastapi-limiter==0.2.0
filetype==1.2.0
flower==2.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-genai==1.70.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.3.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
humanize==4.15.0
idna==3.11
iniconfig==2.3.0
jsonpatch==1.33
jsonpointer==3.1.1
kombu==5.6.2
langchain==1.2.15
langchain-core==1.2.26
langchain-google-genai==4.2.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.2.6
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
pgvector==0.4.2
pinecone==8.1.2
pinecone-plugin-assistant==3.0.3
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyGithub==2.9.0
Pygments==2.20.0
PyJWT==2.12.1
PyNaCl==1.6.2
pyparsing==3.3.2
pyrate-limiter==4.1.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
requests==2.33.1
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
structlog==25.5.0
tenacity==9.1.4
tomli==2.4.1
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.43.0
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
